backend/requirements.txt
# Rural Legal Saathi - Backend Dependencies

# Flask framework
Flask==3.0.0
Flask-CORS==4.0.0

# Environment variables
python-dotenv==1.0.0

# Database - MongoDB
pymongo==4.6.0

# Authentication
JWT==1.3.1
PyJWT==2.11.0
bcrypt==4.1.1

# File handling
Werkzeug==3.0.1

# HTTP requests (for external APIs)
requests==2.31.0

# JSON handling
simplejson==3.19.2

# For future AI integration
# openai==1.6.1
# google-generativeai==0.3.2

# For future OCR integration
# pytesseract==0.3.10
# Pillow==10.1.0
# pdf2image==1.16.3

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production server
gunicorn==21.2.0
